+++ packages/pds/src/index.ts
@@ -17,6 +17,7 @@ import {
   RedisRateLimiter,
   ResponseType,
   XRPCError,
+  createServer,
 } from '@atproto/xrpc-server'
 import apiRoutes from './api'
 import * as authRoutes from './auth-routes'
@@ -24,32 +25,33 @@ import * as basicRoutes from './basic-routes'
 import { ServerConfig, ServerSecrets } from './config'
 import { AppContext, AppContextOptions } from './context'
 import * as error from './error'
-import { createServer } from './lexicon'
-import * as AppBskyFeedGetFeedSkeleton from './lexicon/types/app/bsky/feed/getFeedSkeleton'
+import { app } from './lexicons.js'
 import { httpLogger, loggerMiddleware } from './logger'
 import { proxyHandler } from './pipethrough'
 import compression from './util/compression'
 import { BUILD_HASH, BUILD_TIME } from './version'
 import * as wellKnown from './well-known'
 
+export * from './lexicons.js'
 export { createSecretKeyObject } from './auth-verifier'
 export * from './config'
 export { AppContext } from './context'
 export { Database } from './db'
 export { DiskBlobStore } from './disk-blobstore'
-export { createServer as createLexiconServer } from './lexicon'
 export { httpLogger } from './logger'
 export { type CommitDataWithOps, type PreparedWrite } from './repo'
 export * as repoPrepare from './repo/prepare'
 export { scripts } from './scripts'
 export * as sequencer from './sequencer'
 
-// Legacy export for backwards compatibility
+/**
+ * @deprecated Legacy export for backwards compatibility
+ */
 export type SkeletonHandler = MethodHandler<
   void,
-  AppBskyFeedGetFeedSkeleton.QueryParams,
-  AppBskyFeedGetFeedSkeleton.HandlerInput,
-  AppBskyFeedGetFeedSkeleton.HandlerOutput
+  app.bsky.feed.getFeedSkeleton.$Params,
+  void,
+  app.bsky.feed.getFeedSkeleton.$Output
 >
 
 export class PDS {
@@ -130,8 +132,8 @@ export class PDS {
 
     const { rateLimits } = ctx.cfg
 
-    const server = createServer({
-      validateResponse: false,
+    const server = createServer([], {
+      validateResponse: cfg.service.devMode,
       payload: {
         jsonLimit: 150 * 1024, // 150kb
         textLimit: 100 * 1024, // 100kb
@@ -218,7 +220,7 @@ export class PDS {
     app.use(cors({ maxAge: DAY / SECOND }))
     app.use(basicRoutes.createRouter(ctx))
     app.use(wellKnown.createRouter(ctx))
-    app.use(server.xrpc.router)
+    app.use(server.router)
     app.use(error.handler)
 
     return new PDS({